App_starter/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.7.4,<3.0.0
pydantic-settings>=2.4.0,<3.0.0
supabase==2.9.0
httpx==0.27.0
python-dotenv>=1.0.1,<2.0.0
python-jose[crypto]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.1
email-validator==2.1.0
python-multipart==0.0.6

# Email sending
sendgrid==6.11.0
mailgun==0.1.1
premailer==3.10.0
jinja2>=3.1.4,<4.0.0

# CSV processing
pandas==2.2.0
aiofiles==23.2.1

# Background tasks
celery==5.3.4
aioredis==2.0.1
